refactor: created models for Lab, Booking
